+++ source/mysql/test/common.d
@@ -17,9 +17,9 @@ import std.string;
 import std.traits;
 import std.variant;
 
+import mysql.commands;
 import mysql.common;
 import mysql.connection;
-import mysql.protocol.commands;
 import mysql.protocol.extra_types;
 import mysql.result;
 /+