+++ source/mysql/test/regression.d
@@ -19,9 +19,9 @@ import std.string;
 import std.traits;
 import std.variant;
 
+import mysql.commands;
 import mysql.common;
 import mysql.connection;
-import mysql.protocol.commands;
 import mysql.result;
 import mysql.test.common;